// src/enums/occurrence-types.enum.ts
export enum OccurrenceTypes {
  CREATE_OCCURRENCE_FAILURE = 'occurrences/CREATE_OCCURRENCE_FAILURE',
  CREATE_OCCURRENCE_REQUEST = 'occurrences/CREATE_OCCURRENCE_REQUEST',
  CREATE_OCCURRENCE_SUCCESS = 'occurrences/CREATE_OCCURRENCE_SUCCESS',

  LIST_OCCURRENCES_FAILURE = 'occurrences/LIST_OCCURRENCES_FAILURE',
  LIST_OCCURRENCES_REQUEST = 'occurrences/LIST_OCCURRENCES_REQUEST',
  LIST_OCCURRENCES_SUCCESS = 'occurrences/LIST_OCCURRENCES_SUCCESS',
}